services/chatbot-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
motor==3.5.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.2.0
google-generativeai>=0.8.0
